+++ cmake/CompilerOptions.cmake
@@ -19,7 +19,7 @@ function(enable_project_options target_name)
   endif()
 
   if(MSVC)
-    target_compile_options(${target_name} INTERFACE /utf-8 /Zm10 /EHsc)
+    target_compile_options(${target_name} INTERFACE /utf-8 /EHsc)
   else()
     option(ENABLE_COVERAGE "Enable coverage reporting for gcc/clang" FALSE)
     if(ENABLE_COVERAGE)